+++ jamopp.tests/src/main/java/tools/mdsd/jamopp/test/performance/PerformanceData.java
@@ -19,7 +19,8 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-
+import java.util.List;
+import org.apache.commons.math4.legacy.stat.descriptive.SummaryStatistics;
 import com.google.gson.Gson;
 
 /**
@@ -27,13 +28,28 @@ import com.google.gson.Gson;
  */
 public class PerformanceData {
 	private ArrayList<PerformanceDataPoint> points = new ArrayList<>();
+	private ArrayList<StoragePerformance> storage = new ArrayList<>();
 	
-	public ArrayList<PerformanceDataPoint> getPoints() {
-		return points;
+	public List<PerformanceDataPoint> getPoints() {
+		return (List<PerformanceDataPoint>) points.clone();
+	}
+	
+	public void addPoint(PerformanceDataPoint newPoint) {
+		this.points.add(newPoint);
 	}
 
-	public void setPoints(ArrayList<PerformanceDataPoint> points) {
-		this.points = points;
+	public void setPoints(List<PerformanceDataPoint> points) {
+		this.points.clear();
+		this.points.addAll(points);
+	}
+	
+	public List<StoragePerformance> getStorage() {
+		return (List<StoragePerformance>) storage.clone();
+	}
+	
+	public void setStorage(List<StoragePerformance> storage) {
+		this.storage.clear();
+		this.storage.addAll(storage);
 	}
 	
 	public double getAverageParseTime() {
@@ -42,6 +58,26 @@ public class PerformanceData {
 	
 	public double getAverageResolutionTime() {
 		return (double) points.stream().mapToLong(p -> p.getResolutionTime()).sum() / points.size();
+	}
+	
+	public double getAverageRecoveryTime() {
+		return (double) points.stream().mapToLong(p -> p.getRecoverTime()).sum() / points.size();
+	}
+	
+	public SummaryStatistics getStatistics() {
+		SummaryStatistics stats = new SummaryStatistics();
+		points.forEach(p -> {
+			stats.addValue(p.getParseTime() + p.getResolutionTime() + p.getRecoverTime());
+		});
+		return stats;
+	}
+	
+	public SummaryStatistics getStatisticsWithoutResolution() {
+		SummaryStatistics stats = new SummaryStatistics();
+		points.forEach(p -> {
+			stats.addValue(p.getParseTime() + p.getRecoverTime());
+		});
+		return stats;
 	}
 	
 	public static PerformanceData load(Path file) {